스피드 float으로 수정

+++ Assets/Scripts/Item/ItemData.cs
@@ -47,7 +47,7 @@ public class ItemData : ScriptableObject
     public string AttackSpeed;
     public string Attribute;
     public int attackCountLimitValue;
-    public int realSpeed;
+    public float realSpeed;
     public int speedValue;
     public int accuracyValue;
     public int avoidanceValue;
